Moved AnnotationTypeToString to util/to_string.h
