+++ framework/decode/json_writer.cpp
@@ -34,37 +34,6 @@ GFXRECON_BEGIN_NAMESPACE(decode)
 
 using namespace util;
 
-namespace
-{
-/// @brief Convert an annotation to its string representation
-/// @todo This should be moved into util/to_string.h/cpp after the existing
-/// code there that is tied to the deprecated toJSON paths is eliminated.
-std::string AnnotationTypeToString(const format::AnnotationType& type)
-{
-    std::string str;
-    switch (type)
-    {
-        case format::AnnotationType::kUnknown:
-            str.assign("kUnknown");
-            break;
-        case format::AnnotationType::kText:
-            str.assign("kText");
-            break;
-        case format::AnnotationType::kJson:
-            str.assign("kJson");
-            break;
-        case format::AnnotationType::kXml:
-            str.assign("kXml");
-            break;
-        default:
-            str.assign("OUT_OF_RANGE_ERROR");
-            GFXRECON_LOG_WARNING("format::AnnotationType with out of range value: %lu",
-                                 static_cast<long unsigned>(type));
-            break;
-    }
-    return str;
-}
-} // namespace
 
 JsonWriter::JsonWriter(const JsonOptions&     options,
                        const std::string_view gfxrVersion,
